+++ packages/backend/src/@types/knex-tables.d.ts
@@ -84,7 +84,12 @@ import {
 } from '../database/entities/savedCharts';
 import { SessionTable, SessionTableName } from '../database/entities/sessions';
 import { ShareTable, ShareTableName } from '../database/entities/share';
-import { SpaceTable, SpaceTableName } from '../database/entities/spaces';
+import {
+    SpaceShareTable,
+    SpaceShareTableName,
+    SpaceTable,
+    SpaceTableName,
+} from '../database/entities/spaces';
 import { UserTable, UserTableName } from '../database/entities/users';
 import {
     WarehouseCredentialTable,
@@ -127,5 +132,6 @@ declare module 'knex/types/tables' {
         [ProjectMembershipsTableName]: ProjectMembershipsTable;
         [DbtCloudIntegrationsTableName]: DbtCloudIntegrationsTable;
         [ShareTableName]: ShareTable;
+        [SpaceShareTableName]: SpaceShareTable;
     }
 }