+++ packages/common/src/types/projectMemberProfile.ts
@@ -1,3 +1,6 @@
+import assertUnreachable from '../utils/assertUnreachable';
+import { OrganizationMemberRole } from './organizationMemberProfile';
+
 export enum ProjectMemberRole {
     VIEWER = 'viewer',
     EDITOR = 'editor',