+++ packages/frontend/src/components/ChartConfigPanel/Legend/index.tsx
@@ -3,7 +3,7 @@ import { EchartsLegend, friendlyName } from '@lightdash/common';
 import React, { FC } from 'react';
 import { useForm } from 'react-hook-form';
 import { useVisualizationContext } from '../../LightdashVisualization/VisualizationProvider';
-import BooleanSwitch from '../../ReactHookForm/BooleanSwitch';
+import Checkbox from '../../ReactHookForm/Checkbox';
 import Form from '../../ReactHookForm/Form';
 import Input from '../../ReactHookForm/Input';
 import Select from '../../ReactHookForm/Select';
@@ -38,11 +38,12 @@ const LegendPanel: FC = () => {
             onSubmit={() => undefined}
             onBlur={handleSubmit(setLegend)}
         >
-            <BooleanSwitch
+            <Checkbox
                 name="show"
-                label="Show legend"
+                checkboxProps={{ label: 'Show legend' }}
                 defaultValue={showDefault}
             />
+
             <Collapse
                 isOpen={
                     dirtyEchartsConfig?.legend