+++ packages/frontend/src/components/Explorer/FiltersCard/FiltersCard.tsx
@@ -175,7 +175,7 @@ const FiltersCard: FC = memo(() => {
                             interactionKind="hover"
                             placement={'bottom-start'}
                         >
-                            <Tag large round minimal>
+                            <Tag round minimal>
                                 {totalActiveFilters} active filter
                                 {totalActiveFilters === 1 ? '' : 's'}
                             </Tag>